requirements.txt
rumps==0.4.0
pynput==1.7.6
google-genai>=1.47.0
python-dotenv==1.0.0
pyperclip==1.8.2
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
